requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.16
django-allauth==0.51.0
django-ckeditor==6.3.2
django-environ==0.8.1
django-js-asset==2.0.0
django-model-utils==4.3.1
django-mptt==0.14.0
django-notifications-hq==1.7.0
django-sslserver==0.22
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
jsonfield==3.1.0
Markdown==3.4.1
mysqlclient==1.4.6
oauthlib==3.2.2
Pillow==9.4.0
pip==22.3.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2021.3
requests==2.28.2
requests-oauthlib==1.3.1
setuptools==65.6.3
sqlparse==0.4.2
swapper==1.3.0
toml==0.10.2
typing_extensions==4.4.0
urllib3==1.26.14
wheel==0.37.1
whitenoise==5.3.0
wincertstore==0.2
zipp==3.13.0
